requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
whitenoise==6.6.0
gunicorn==21.2.0
